Support Sonnet 3.7
