+++ apps/studio/electron/main/chat/index.ts
@@ -1,5 +1,6 @@
 import { PromptProvider } from '@onlook/ai/src/prompt/provider';
 import { listFilesTool, readFileTool } from '@onlook/ai/src/tools';
+import { CLAUDE_MODELS, LLMProvider } from '@onlook/models';
 import {
     ChatSuggestionSchema,
     StreamRequestType,
@@ -11,7 +12,7 @@ import { MainChannels } from '@onlook/models/constants';
 import { generateObject, streamText, type CoreMessage, type CoreSystemMessage } from 'ai';
 import { mainWindow } from '..';
 import { PersistentStorage } from '../storage';
-import { CLAUDE_MODELS, initModel, LLMProvider } from './llmProvider';
+import { initModel } from './llmProvider';
 
 class LlmManager {
     private static instance: LlmManager;
