+++ packages/models/src/llm/index.ts
@@ -1,0 +1,8 @@
+export enum LLMProvider {
+    ANTHROPIC = 'anthropic',
+}
+
+export enum CLAUDE_MODELS {
+    SONNET = 'claude-3-5-sonnet-20241022',
+    HAIKU = 'claude-3-5-haiku-20241022',
+}
